Clear cache when changing repo
This won't happen in a normal CLI run, but in the tests we render two
different repos.

Although, if two blobs have the same OID (SHA1) then (the working
assumption is that) they have the same contents.  If they also have
the smae entry name, they should be highlighted the same anyway.  But
it seems cleaner to tie the results to the repo.